# api/openapi.yaml
openapi: 3.1.0
info:
  title: Zyvor OTA local operator API
  version: 0.1.0
  description: >-
    Served over a mode-0600 Unix socket. Filesystem permissions authorize the
    operator. Loopback HTTP is available only with the simulator backend.
servers:
  - url: http://unix
paths:
  /v1/status:
    get:
      operationId: status
      responses:
        '200':
          description: Agent version, device ID, active job, accepted sequence and event counters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /v1/jobs/{id}:
    get:
      operationId: getJob
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retained job, signed assignment, release and current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/assignments:
    post:
      operationId: submitAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assignment'
      responses:
        '202':
          description: Accepted or identical retained job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid JSON or body exceeds 256 KiB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Invalid signature/policy, replay, conflict or busy engine; retry unchanged assignment when busy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/events:
    get:
      operationId: getEvents
      responses:
        '200':
          description: Ordered unacknowledged events; empty outbox can serialize as null
          content:
            application/json:
              schema:
                type: [array, 'null']
                items:
                  $ref: '#/components/schemas/Event'
  /v1/events/ack:
    post:
      operationId: acknowledgeEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sequence]
              properties:
                sequence:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Acknowledged contiguous events through sequence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Future acknowledgment or persistence error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/reboot:
    post:
      operationId: requestReboot
      responses:
        '200':
          description: Reboot intent persisted and request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '409':
          description: Wrong state, outside window, busy backend or denied reboot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/recover-abort:
    post:
      operationId: recoverAbort
      responses:
        '200':
          description: Ambiguous target marked bad and previous slot restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '409':
          description: Requires NeedsRecovery, idle backend and original booted slot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/gc:
    post:
      operationId: cleanCache
      responses:
        '200':
          description: Cache removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '409':
          description: Active job, busy engine or storage error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /metrics:
    get:
      operationId: metrics
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    Envelope:
      type: object
      additionalProperties: false
      required: [key_id, payload, signature]
      properties:
        key_id:
          type: string
        payload:
          type: string
          contentEncoding: base64
          description: Exact JSON bytes of the release.schema.json document
        signature:
          type: string
          contentEncoding: base64
          description: Ed25519 signature over those exact payload bytes
    Assignment:
      type: object
      additionalProperties: false
      required: [job_id, device_id, release, not_before, deadline]
      properties:
        job_id:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9._-]{0,95}$'
        device_id:
          type: string
        release:
          $ref: '#/components/schemas/Envelope'
        not_before:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        auto_reboot:
          type: boolean
          default: false
        campaign_id:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    Ok:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
    Status:
      type: object
      required: [version, device_id, backend, high_sequence, event_sequence, pending_events]
      properties:
        version:
          type: string
        device_id:
          type: string
        backend:
          type: string
        active:
          $ref: '#/components/schemas/Job'
        high_sequence:
          type: integer
        event_sequence:
          type: integer
        pending_events:
          type: integer
    Job:
      type: object
      required: [assignment, release, fingerprint, state, updated]
      properties:
        assignment:
          $ref: '#/components/schemas/Assignment'
        release:
          type: object
        fingerprint:
          type: string
        state:
          type: string
        error:
          type: string
        updated:
          type: string
          format: date-time
    Event:
      type: object
      required: [sequence, job_id, state, time]
      properties:
        sequence:
          type: integer
          minimum: 1
        job_id:
          type: string
        state:
          type: string
          enum: [accepted, downloading, verified, installing, awaiting_reboot, checking_health, committing, committed, rollback_pending, rolled_back, failed, needs_recovery]
        error:
          type: string
        time:
          type: string
          format: date-time